+++ muffinscript/parser.py
@@ -48,6 +48,9 @@ def _parse_expression(
 
         if (
             (isinstance(token, str) and token.startswith('"') and token.endswith('"'))
+            or token == "true"  # nosec
+            or token == "false"  # nosec
+            or token == "null"  # nosec
             or isinstance(token, int)
             or isinstance(token, float)
         ):