+++ test/unit/test_lexer.py
@@ -33,17 +33,26 @@ def test_tokenizer():
     tokens = tokenize("foo = 2.5", 9)
     assert tokens == ["foo", "=", 2.5]
 
-    with pytest.raises(MuffinScriptSyntaxError) as error:
-        tokenize("foo = .test", 10)
-    assert str(error.value) == "\033[31mERROR\033[0m - Invalid float on line 10"
+    tokens = tokenize("foo = true", 10)
+    assert tokens == ["foo", "=", "true"]
+
+    tokens = tokenize("foo = false", 11)
+    assert tokens == ["foo", "=", "false"]
+
+    tokens = tokenize("foo = null", 12)
+    assert tokens == ["foo", "=", "null"]
 
     with pytest.raises(MuffinScriptSyntaxError) as error:
-        tokenize("foo = 2.3.4", 11)
-    assert str(error.value) == "\033[31mERROR\033[0m - Invalid float on line 11"
+        tokenize("foo = .test", 13)
+    assert str(error.value) == "\033[31mERROR\033[0m - Invalid float on line 13"
 
     with pytest.raises(MuffinScriptSyntaxError) as error:
-        tokenize("?", 9)
-    assert str(error.value) == "\033[31mERROR\033[0m - Unknown token on line 9: ?"
+        tokenize("foo = 2.3.4", 14)
+    assert str(error.value) == "\033[31mERROR\033[0m - Invalid float on line 14"
+
+    with pytest.raises(MuffinScriptSyntaxError) as error:
+        tokenize("?", 15)
+    assert str(error.value) == "\033[31mERROR\033[0m - Unknown token on line 15: ?"
 
     with pytest.raises(MuffinScriptSyntaxError) as error:
         tokenize('p("hello world)', 10)