+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next Release
 
-- Data types: `str`, `int`, `float`
+- Data types: `str`, `int`, `float`, `bool`, `null`
 - Simple arithmetic
 - Variable assignment
 - Print to console